pyproject.toml
[tool.poetry]
name = "art-code-to-sheets"
version = "0.1.0"
description = "daily code assistant to create a google sheet from the daily schedule and then add cpt codes and modifiers or units"
authors = ["nate ginn <growyourbiz4ever@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tabula-py = "^2.10.0"
pandas = "^2.2.3"
google-auth = "^2.36.0"
google-api-python-client = "^2.154.0"
pytest = "^8.3.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
